refactor: extract issue spec cli module
